bk 0620 게시판 crud 완성!

+++ src/main/java/com/green/nowon/dto/MymemoDTO.java
@@ -24,7 +24,7 @@ public class MymemoDTO {
 	private LocalDateTime createdDate;
 	private LocalDateTime updatedDate;
 	
-	public MymemoDTO mno(int mno) {
+	public MymemoDTO mno(long mno) {
 		this.mno=mno;
 		return this;
 	}
